Implement custom assertion libraries.
* Ability to use chai, expect.js, should.js or built-in assertions
* Create separate test files for each assertion style
