+++ .gitignore
@@ -1,3 +1,5 @@
+.idea
 .DS_Store
 
-/node_modules/
+node_modules
+bower_components
