+++ README.md
@@ -14,7 +14,7 @@ module.exports = function(config) {
     // list of files / patterns to load in the browser
     files: [
       './should.js',
-      './testSpec.js',
+      './test.spec.js',
     ],
     exclude: [
     ],
