+++ bower.json
@@ -1,0 +1,28 @@
+{
+  "name": "karma-e2e-dsl",
+  "main": "index.js",
+  "version": "1.0.0",
+  "homepage": "https://github.com/ducksoupdev/karma-e2e-dsl",
+  "authors": [
+    "Matt Levy <ducksoupdev@gmail.com>"
+  ],
+  "moduleType": [
+    "node"
+  ],
+  "license": "MIT",
+  "private": true,
+  "ignore": [
+    "**/.*",
+    "node_modules",
+    "bower_components",
+    "test",
+    "tests"
+  ],
+  "dependencies": {
+    "jquery": "~2.1.3",
+    "underscore": "~1.7.0",
+    "should": "~4.6.2",
+    "expect": "~0.3.1",
+    "chai": "~1.10.0"
+  }
+}
