+++ index.js
@@ -4,8 +4,8 @@ var createPattern = function(path) {
 
 var initDsl = function(files) {
   files.unshift(createPattern(__dirname + '/dist/karma-e2e-dsl.min.js'));
-  files.unshift(createPattern(__dirname + '/jquery.min.js'));
-  files.unshift(createPattern(__dirname + '/underscore.min.js'));
+  files.unshift(createPattern(__dirname + '/bower_components/jquery/dist/jquery.min.js'));
+  files.unshift(createPattern(__dirname + '/bower_components/underscore/underscore-min.js'));
 };
 
 initDsl.$inject = ['config.files'];
