--- karma-ng.conf.js
@@ -1,79 +1,0 @@
-// Karma configuration
-// Generated on Thu Oct 31 2013 11:36:56 GMT+0800 (CST)
-
-module.exports = function(config) {
-  config.set({
-
-    // base path, that will be used to resolve files and exclude
-    basePath: './',
-
-
-    // frameworks to use
-    frameworks: ['ng-scenario'],
-
-
-    // list of files / patterns to load in the browser
-    files: [
-      './testSpec.js'
-    ],
-
-
-    // list of files to exclude
-    exclude: [
-      
-    ],
-
-    proxies: {
-        '/app': 'http://localhost:8000/',
-        '/weibo': 'http://www.weibo.com'
-    },
-
-
-    // test results reporter to use
-    // possible values: 'dots', 'progress', 'junit', 'growl', 'coverage'
-    reporters: ['progress'],
-
-
-    // web server port
-    port: 9876,
-
-
-    // enable / disable colors in the output (reporters and logs)
-    colors: true,
-
-
-    // level of logging
-    // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
-    logLevel: config.LOG_INFO,
-
-
-    // enable / disable watching file and executing tests whenever any file changes
-    autoWatch: true,
-
-
-    // Start these browsers, currently available:
-    // - Chrome
-    // - ChromeCanary
-    // - Firefox
-    // - Opera (has to be installed with `npm install karma-opera-launcher`)
-    // - Safari (only Mac; has to be installed with `npm install karma-safari-launcher`)
-    // - PhantomJS
-    // - IE (only Windows; has to be installed with `npm install karma-ie-launcher`)
-    browsers: ['Chrome', 'Firefox'],
-
-
-    // If browser does not capture in given timeout [ms], kill it
-    captureTimeout: 60000,
-
-
-    // Continuous Integration mode
-    // if true, it capture browsers, run tests and exit
-    singleRun: false,
-
-    plugins : [
-      'karma-ng-scenario',
-      'karma-chrome-launcher',
-      'karma-firefox-launcher'
-    ]
-  });
-};
