+++ karma.assert.conf.js
@@ -7,9 +7,10 @@ module.exports = function(config) {
         files: [
             './bower_components/underscore/underscore-min.js',
             './bower_components/jquery/dist/jquery.min.js',
+            './bower_components/chai/chai.js',
             './expectations.js',
             './karma-e2e-dsl.js',
-            './test.spec.js'
+            './test-assert.spec.js'
         ]
     });
 };
