+++ karma.should.conf.js
@@ -7,9 +7,10 @@ module.exports = function(config) {
         files: [
             './bower_components/underscore/underscore-min.js',
             './bower_components/jquery/dist/jquery.min.js',
+            './bower_components/should/should.min.js',
             './expectations.js',
             './karma-e2e-dsl.js',
-            './test.spec.js'
+            './test-should.spec.js'
         ]
     });
 };
